have condtion guard check length of args and no args, have condition at point of json creation check if file exists, ignore meta.json, have create_json fn called if pkg installation is accepted and successful

+++ extract.sh
@@ -1,8 +1,9 @@
 #! /usr/bin/bash
 
-function get_value() {
-  LINES="$#"
+ARG_QN=$#
 
+
+function get_value() {
   for LINE in $1; do
     IFS='=' read -ra PARTS <<< "$LINE"
 
@@ -18,16 +19,32 @@ function get_value() {
 }
 
 function create_json() {
-  ENCODED=$(exiftool -a -r -G1 -s $1 | awk '/Certificate/ {print $4}')
-  echo $ENCODED | base64 -d > meta.json
+  if test -f "./meta.json"; then
+    echo "meta.json exists"
+  else
+    ENCODED=$(exiftool -a -r -G1 -s $1 | awk '/Certificate/ {print $4}')
+    echo $ENCODED | base64 -d > meta.json
+  fi
 }
 
 IFS=$'\n'
 PROFILE=$(cat "/etc/lsb-release")
 TARGET="DISTRIB_DESCRIPTION"
 OS=$(get_value "${PROFILE[@]}" "$TARGET")
-if [[ "$OS" == "Manjaro" || "$OS" == "Arch" ]]
-then
+
+if [ $# -gt 1 ]; then
+  echo "Only one argument is required. $ARG_QN was given"
+
+  exit 2
+fi
+
+if [ -z $1 ]; then
+  echo 'Missing location of image as argument.'
+  
+  exit 2
+fi
+
+if [[ "$OS" == "Manjaro" || "$OS" == "Arch" ]]; then
   {
     exiftool -ver && {
       echo 'continue'
@@ -36,8 +53,8 @@ then
     }
   } || {
     echo 'download exiftool'
-    pacman -S "perl-image-exiftool"
-
-    create_json $1
+    pacman -S "perl-image-exiftool" && create_json $1
   }
+else
+  echo 'OS must be Arch linux or Manjaro'
 fi
